Add basic unit tests
