+++ rpg/__init__.py
@@ -1,0 +1,3 @@
+from .utils import *
+from .factories import *
+from .example import RpgExampleComparativePygame
