+++ tests/__init__.py
@@ -1,0 +1,1 @@
+import sys, os; sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
