+++ tests/test_goap_agent.py
@@ -1,0 +1,9 @@
+from goap.agent import Agent
+
+
+def test_agent_defaults():
+    agent = Agent(name="Tester")
+    assert agent.Name == "Tester"
+    assert agent.Goals == []
+    assert agent.Actions == []
+    assert agent.Sensors == []
