+++ tests/test_goap_init.py
@@ -1,0 +1,6 @@
+import goap
+
+
+def test_package_imports():
+    assert hasattr(goap, 'Action')
+    assert hasattr(goap, 'Agent')
