+++ tests/test_goap_planning.py
@@ -1,0 +1,10 @@
+from goap.actions import Action
+from goap.planning import ActionNode
+
+
+def test_action_node_equality():
+    act1 = Action(name="test")
+    node1 = ActionNode(act1, state={}, parameters={})
+    node2 = ActionNode(act1, state={}, parameters={})
+    assert node1 == node2
+    assert hash(node1) == hash(node2)
