+++ tests/test_goap_types.py
@@ -1,0 +1,5 @@
+from typing import get_origin
+from goap.types import StateDictionary
+
+def test_state_dictionary_alias():
+    assert get_origin(StateDictionary) is dict
