+++ tests/test_goap_utils.py
@@ -1,0 +1,15 @@
+from goap.utils import PriorityQueue, PriorityQueueNode
+
+class Node(PriorityQueueNode):
+    pass
+
+def test_priority_queue():
+    pq = PriorityQueue()
+    a = Node()
+    b = Node()
+    pq.enqueue(a, 2)
+    pq.enqueue(b, 1)
+    first = pq.dequeue()
+    second = pq.dequeue()
+    assert first is b
+    assert second is a
