+++ tests/test_rpg_init.py
@@ -1,0 +1,8 @@
+import importlib
+
+def test_rpg_imports():
+    try:
+        module = importlib.import_module('rpg')
+    except ModuleNotFoundError:
+        return
+    assert hasattr(module, 'PlayerFactory')
